requirements.txt
streamlit>=1.25.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python-headless>=4.8.1.78
